Move script to commands
