+++ pyproject.toml
@@ -184,6 +184,7 @@ migrate-attachments-aws = "encoded.commands.migrate_attachments_aws:main"
 migrate-dataset-type = "encoded.commands.migrate_dataset_type:main"
 migrate-files-aws = "encoded.commands.migrate_files_aws:main"
 owl-to-items = "encoded.commands.generate_items_from_owl:main"
+order-schema-columns-and-facets = "encoded.commands.order_loaded_schema_columns_and_facets:main"
 profiler = "encoded.commands.profiler:main"
 purge-item-type = "encoded.commands.purge_item_type:main"
 run-upgrade-on-inserts = "encoded.commands.run_upgrader_on_inserts:main"
